58.2 - Update case (read/create), project, & contact page data display & added buttons, fragments & block for modal.

+++ src/main/java/com/jamersc/springboot/financialhub/model/Cases.java
@@ -84,14 +84,14 @@ public class Cases {
     private Date endDate;
 
     @Column(name = "created_by")
-    private int createdBy;
+    private Long createdBy;
 
     @Column(name = "created_at", updatable = false)
     @Temporal(TemporalType.TIMESTAMP)
     private Date createdAt;
 
     @Column(name = "updated_by")
-    private int updatedBy;
+    private Long updatedBy;
 
     //@Column(name = "updated_at")
     @Temporal(TemporalType.TIMESTAMP)
